refactor: move UserDir
