+++ src/MeoAssistant/Utils/Logger.hpp
@@ -13,6 +13,7 @@
 #include "AsstUtils.hpp"
 #include "Meta.hpp"
 #include "SingletonHolder.hpp"
+#include "UserDir.hpp"
 #include "Version.h"
 
 #if defined(__APPLE__) || defined(__linux__)
@@ -415,7 +416,7 @@ namespace asst
     private:
         friend class SingletonHolder<Logger>;
 
-        Logger() : m_directory(utils::UserDir::get_instance().get())
+        Logger() : m_directory(UserDir::get_instance().get())
         {
             check_filesize_and_remove();
             log_init_info();
